+++ aom_dsp/grain_synthesis.h
@@ -226,7 +226,7 @@ void generate_grain_uv_c(grain_values *grain_data,
 void init_scaling_function_extern(const int scaling_points[][2], int num_points,
                                   int scaling_lut[]);
 
-int generate_grain_image(const aom_film_grain_t *params, int *luma, int *cb, int *cr, int height, int width,
+int generate_grain_image(const aom_film_grain_t *params, uint8_t *luma, uint8_t *cb, uint8_t *cr, int height, int width,
                            int luma_stride, int chroma_stride, int chroma_subsamp_y,
                            int chroma_subsamp_x, int mc_identity);
